[api] Move api -> api/strava
