+++ api/strava/service.go
@@ -1,4 +1,4 @@
-package api
+package strava
 
 // Copied from https://github.com/strava/go.strava/blob/99ebe972ba16ef3e1b1e5f62003dae3ac06f3adb/service.go
 // as dependency for current_athlete.go
